% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ces0406}
\alias{ces0406}
\title{2004-2006 Canadian Election Study Surveys.}
\format{
A data frame with 1019 variables.

For further details, see
}
\usage{
ces0406
}
\description{
The combination of datasets from the 2004, 2006, and 2008 Canadian Election Study (CES).
}
\keyword{datasets}
