% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ces0411}
\alias{ces0411}
\title{2004, 2006, 2008, 2011 Canadian Election Study Surveys.}
\format{
A data frame with 2135 variables.

For further details, see \url{https://ces-eec.arts.ubc.ca/english-section/surveys/}
}
\usage{
ces0411
}
\description{
The combination of datasets from the 2004, 2006, 2008, and 2011 Canadian Election Study (CES).
}
\keyword{datasets}
