% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ces1965}
\alias{ces1965}
\title{1965 November Canadian Election Study Survey.}
\format{
A data frame with 350 variables.

For further details, see
}
\usage{
ces1965
}
\description{
The dataset from the 1965 Canadian Election Study (CES).
}
\keyword{datasets}
