% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ces1968}
\alias{ces1968}
\title{1968 November Canadian Election Study Survey.}
\format{
A data frame with 457 variables.

For further details, see
}
\usage{
ces1968
}
\description{
The dataset from the 1968 Canadian Election Study (CES).
}
\keyword{datasets}
