% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ces1974}
\alias{ces1974}
\title{1974 Canadian Election Study Survey.}
\format{
A data frame with 481 variables.

For further details, see
}
\usage{
ces1974
}
\description{
The dataset from the 1974 Canadian Election Study (CES).
}
\keyword{datasets}
