% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ces1984}
\alias{ces1984}
\title{1984 Canadian Election Study Survey.}
\format{
A data frame with 641 variables.

For further details, see
}
\usage{
ces1984
}
\description{
The dataset from the 1984 Canadian Election Study (CES).
}
\keyword{datasets}
