% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ces1988}
\alias{ces1988}
\title{1988 Canadian Election Study Survey.}
\format{
A data frame with 548 variables.

For further details, see
}
\usage{
ces1988
}
\description{
The dataset from the 1988 Canadian Election Study (CES).
}
\keyword{datasets}
