% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ces1993}
\alias{ces1993}
\title{1993 Canadian Election Study Survey.}
\format{
A data frame with 872 variables.

For further details, see
}
\usage{
ces1993
}
\description{
The dataset from the 1993 Canadian Election Study (CES).
}
\keyword{datasets}
