% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ces1997}
\alias{ces1997}
\title{1997 Canadian Election Study Survey.}
\format{
A data frame with 626 variables.

For further details, see
}
\usage{
ces1997
}
\description{
The dataset from the 1997 Canadian Election Study (CES).
}
\keyword{datasets}
