% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ces2000}
\alias{ces2000}
\title{2000 Canadian Election Study Survey.}
\format{
A data frame with 780 variables.

For further details, see
}
\usage{
ces2000
}
\description{
The dataset from the 2000 Canadian Election Study (CES).
}
\keyword{datasets}
