% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ces2004}
\alias{ces2004}
\title{2004 Canadian Election Study Survey.}
\format{
A data frame with 632 variables.

For further details, see
}
\usage{
ces2004
}
\description{
The dataset from the 2004 Canadian Election Study (CES).
}
\keyword{datasets}
