% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ces2008}
\alias{ces2008}
\title{2008 Canadian Election Study Survey.}
\format{
A data frame with 531 variables.

For further details, see
}
\usage{
ces2008
}
\description{
The dataset from the 2008 Canadian Election Study (CES).
}
\keyword{datasets}
