% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ces2011}
\alias{ces2011}
\title{2011 Canadian Election Study Survey.}
\format{
A data frame with 596 variables.

For further details, see \url{https://ces-eec.arts.ubc.ca/english-section/surveys/}
Fournier, Patrick, Fred Cutler, Stuart Soroka and Dietlind Stolle. 2011. The 2011 Canadian Election Study. [dataset]
}
\usage{
ces2011
}
\description{
The combination of online and phone survey datasets from the 2015 Canadian Election Study (CES).
}
\keyword{datasets}
