% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ces2015_combo}
\alias{ces2015_combo}
\title{2015 Canadian Election Study Online/Phone Survey.}
\format{
A data frame with 665 variables.

For further details, see \url{https://ces-eec.arts.ubc.ca/english-section/surveys/}
Fournier, Patrick, Fred Cutler, Stuart Soroka and Dietlind Stolle. 2015. The 2015 Canadian Election Study. [dataset]
}
\usage{
ces2015_combo
}
\description{
The combination of online and phone survey datasets from the 2015 Canadian Election Study (CES).
}
\keyword{datasets}
