% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ces2019_phone}
\alias{ces2019_phone}
\title{2019 Canadian Election Study Phone Survey.}
\format{
A data frame with 273 variables.

For further details, see \url{https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/8RHLG1}
Stephenson, Laura B; Harell, Allison; Rubenson, Daniel; Loewen, Peter John, 2020, "2019 Canadian Election Study - Phone Survey",
\url{https://doi.org/10.7910/DVN/8RHLG1}, Harvard Dataverse, V1, UNF:6:eyR28qaoYlHj9qwPWZmmVQ== [fileUNF]
}
\usage{
ces2019_phone
}
\description{
The phone survey dataset from the 2019 Canadian Election Study (CES).
}
\keyword{datasets}
