% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ces2019_web}
\alias{ces2019_web}
\title{2019 Canadian Election Study Online Survey.}
\format{
A data frame with 620 variables.

For further details, see \url{https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/DUS88V}
Stephenson, Laura B; Harell, Allison; Rubenson, Daniel; Loewen, Peter John, 2020, "2019 Canadian Election Study - Online Survey",
\url{https://doi.org/10.7910/DVN/DUS88V}, Harvard Dataverse, V1
}
\usage{
ces2019_web
}
\description{
The online survey dataset from the 2019 Canadian Election Study (CES).
}
\keyword{datasets}
