% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ces72_jnjl}
\alias{ces72_jnjl}
\title{1972 Juny/July Canadian Election Study Survey.}
\format{
A data frame with 137 variables.

For further details, see
}
\usage{
ces72_jnjl
}
\description{
The dataset from the 1972 June and July Canadian Election Study (CES).
}
\keyword{datasets}
