% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ces72_nov}
\alias{ces72_nov}
\title{1972 November Canadian Election Study Survey.}
\format{
A data frame with 188 variables.

For further details, see
}
\usage{
ces72_nov
}
\description{
The dataset from the 1972 November Canadian Election Study (CES).
}
\keyword{datasets}
