% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ces72_sep}
\alias{ces72_sep}
\title{1972 September Canadian Election Study Survey.}
\format{
A data frame with 111 variables.

For further details, see
}
\usage{
ces72_sep
}
\description{
The dataset from the 1972 September Canadian Election Study (CES).
}
\keyword{datasets}
