% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ces7480}
\alias{ces7480}
\title{1974-1980 Canadian Election Study Survey.}
\format{
A data frame with 1259 variables.

For further details, see
}
\usage{
ces7480
}
\description{
The dataset composed of responses from the respondents who answered both the 1974 and 1979 Canadian Election Study (CES).
}
\keyword{datasets}
