\name{cesnames}
\alias{cesnames}
\docType{data}
\title{
Species names translation
}
\description{
A datset containing species names in different languages for use by \code{\link{setceslang}}
}
\usage{data(ukdata)}
\format{
  A data frame with 666 species names in English ad 11 other languages.
}
\keyword{datasets}
