\name{extract.counts}
\alias{extract.counts}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Get species data from a CES datset
}
\description{
Extracts catch data for a particular species from a CES object, usually called through \code{extract.data}.
}
\usage{
extract.counts(cesobj, species=0, age=0, plots=NULL)
}
\arguments{
  \item{cesobj}{
  A dataframe containing CES data usually generated by /code{readces}
}
  \item{species}{
  Species for which to extract data. This should be a standard Euring numeric code, for example Turdus merula would be 11870 (note the last digit, to indicate race, should always be zero). If species=0, summarises catch total across all species.
}
  \item{age}{
  Euring age code 3 for juveniles and 4 for adults, any other value is an error.
}
  \item{plots}{
  An object indicating which sites were covered in each year. Normally this would be generated by \code{extract.coverage}, if it is absent the coverage data will be extracted silently.   
}
}
\value{
  A datatframe containing the number of birds caught at each site in each year. Usually with an identifier for the species to which the data refer.
}
\author{
Rob Robinson
}
\seealso{
\code{\link{extract.data}}
}
\keyword{ manip }
