% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{onlinesurvey}
\alias{onlinesurvey}
\title{2019 Canadian Election Online Survey.}
\format{
A data frame with 620 variables and 37822 observations.

For further details, see \url{https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/DUS88V}
Stephenson, Laura B; Harell, Allison; Rubenson, Daniel; Loewen, Peter John, 2020, "2019 Canadian Election Study - Online Survey",
\url{https://doi.org/10.7910/DVN/DUS88V}, Harvard Dataverse, V1
}
\usage{
onlinesurvey
}
\description{
Both the online and phone survey datasets from the 2019 Canadian Election Survey (CES)
}
\keyword{datasets}
