\name{print.ces}
\alias{print.ces}
\title{
Print methods for CES objects
}
\description{
Largely a placeholder currently, but handles summary tables gracefully }
\usage{
\method{print}{ces}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
A CES data object
}
  \item{...}{
Other arguments, currently they will have effect!
}
}
\details{
Largely a placeholder currently, but handles summary tables gracefully.
}
\author{
Rob Robinson
}
\keyword{ utilities }
