\name{select.data}
\alias{select.data}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Subsets a CES dataset
}
\description{
Select sites from a CES data object, either from a ist of specified sites or habitat.
}
\usage{
select.data(x, sites = 0, habitat = "")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
A CES data object, typically generated by \code{readces()}
}
  \item{sites}{
A vector of site codes
}
  \item{habitat}{
Two-letter code identifying habitat required. Generally the available codes are 
	'RB' Reedbed, 'WS' Wet Scrub, 'DS' Dry/Thorn Scrub and 'WD' Woodland
	but certain countries may have other scheme-specific codes }
}
\details{
An easy way of subsetting a CES data object to select particular sites. Experienced users may prefer to do this directly.
}
\value{
A ces data object}
\author{
Rob Robinson
}
\seealso{
\code{\link{readces}}
}
\examples{
data(ukdata)
sitelist = c(401, 403, 405, 407, 408, 414, 417, 447, 463) # select sites with complete data
complete = select.data(ukdata, sites=sitelist)

scrub = select.data(ukdata, habitat='DS')

}
\keyword{ manip }
