\name{ces.table}
\alias{ces.table}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Create a table of results for a list of species 
}
\description{
Runs a specified set of models (see details) for a list of species and present the results in a table. 
}
\usage{
ces.table(cesobj, species=NA, columns=c("A-1", "P-1", "S-1"), base=100, plots=NULL, min.n=100, min.ch=50, ndigits=2, year=-1, change=FALSE, visit.corr=TRUE)
}
\arguments{
  \item{cesobj}{
  a CES dataframe typically generated by \code{\link{readces}}.
}
  \item{species}{
  a numeric vector of Euring species codes, these can be checked using \code{\link{get.euring}}, if missing use min.n=.
}
  \item{columns}{
  a character vector specifying which models should be applied to each species, see details. 
}
  \item{base}{
  which number should indicate no change: 0, 1, or 100?
}
  \item{plots}{
  an object containing plot coverage data created by \code{\link{extract.coverage}}. If missing one will be created containing all sites and years.
}
  \item{min.n}{
  if species is not specified, include all species with more than this number of captures.
}
  \item{min.ch}{
  for survival models, how many individuals need to be captured for a site to be included, see \code{\link{extract.ch}}.
}
  \item{ndigits}{
  numeric, how many decimal places to round the results to, see note.
}
  \item{year}{
  the reference year for the comparisons if not the last.
}
  \item{change}{
  logical, for comparison models return the change or the relative value (e.g. +0.2 or 1.2)?
}
  \item{visit.corr}{
  logical, include an offset for missed visits in the models, see \code{\link{index}}.
}
}
\details{
Models are specified using three characters: (i) the parameter ('A'dult and 'J'uvenile numbers, 'P'roductivity and 'S'urvival, (ii) the model type, currently either '-' to compare to previous years or '/' for a trend, and (iii) the number of (previous) years over which the comparison or trend should be calculated. Thus, for example. "A-5" would compare the current year's estimate of adult captures to the previous 5 years, see \code{\link{index}} for more details on this. The default is to compare with the previous year ("-1"). 

For change ('-') models the estimates will be either be the index relative to the base value (so -0.15, 0.85, or 85.0 for base = 0, 1, 100 respectively) or the relevant change (if change=TRUE then -0.15, -0.15, 15.0 in those instances). For trend ('/') models the estimate will be a slope (mean annual change), on either the log (abundance) or logit (survival/productivity) scale (and might be expressed as either 0.07 or 7.00, for example, depending on whether or not base=100).

A simple table will be printed with the change/trend estimate (along with confidence limits and an indication of significance) for each model (columns) and species (rows). These numbers are returned as a dataframe, see 'value' below. 
}
\value{
a list with three components: table - a dataframe that mimics the printed table, so in 'wide' format; estimates - a dataframe in 'long' format, one row per entry; results - a named list of the model objects, in the form "T_xxxxx_mmm", where xxxxx is the Euring code number, mmm the model specification.
}
\author{
Rob Robinson
}
\note{
ndigits is strict, the returned 'estimates' dataframe will be rounded to this level as well.
}
\seealso{
\code{\link{index}}
}
\examples{
\dontrun{
# compare adult numbers in the current yaer to the previous year, 
# survival with the previous five years and a 10-year trend in productivity 
results <- ces.table(ukdata, columns=c('A-0', 'S-5', 'P/10'))
}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
