\name{ces.table}
\alias{ces.table}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Creat a table of results for a list of species 
}
\description{
Runs a specified set of models (see details) for a list of species and present the results in a table. 
}
\usage{
ces.table(cesobj, species=NA, columns=c("A-0", "P-0", "S-0"), base=100, plots=NULL, min.n=100, min.ch=50, ndigits=2, year=-1, change=FALSE, visit.corr=TRUE)
}
\arguments{
  \item{cesobj}{
  a dataframe typically generated by \code{\link{readces}}.
}
  \item{species}{
  a numeric vector of Euring species codes, these can be checked using \code{\link{get.euring}}.
}
  \item{columns}{
  a character vector specifying which models should be applied to each species, see details. 
}
  \item{base}{
  which number should indicate no change: 0, 1, or 100?
}
  \item{plots}{
  an object containing plot coverage data created by \code{\link{extract.coverage}}. 
}
  \item{min.n}{
  if species is not specified, include all species with more than this number of captuures.
}
  \item{min.ch}{
  for survival models, how many individuals need to be captured for a site to be included, see \code{\link{extract.ch}}.
}
  \item{ndigits}{
  numeric, how many decimal places to include.
}
  \item{year}{
  which year should be treated as the reference year?
}
  \item{change}{
  logical, for comparison models return the change or the relative value (e.g. +0.2 or 1.2)?
}
  \item{visit.corr}{
  logical, include an offset for missed visits in the models, see \code{\link{index}}.
}
}
\details{
Models are specified using three characters: (i) the parameter ('A'dult and 'J'uvenile numbers, 'P'roductivity and 'S'urvival, (ii) the model type, currently either '-' to compare to previous years or '/' for a trend, and (iii) the number of (previous) years over which the comparison or trend should be calculated. Thus, for example. "A-5" would compare the current year's estimate of adult captures to the previous 5 years, see \code{\link{index}} for more details on this. 

A simple table will be printed with the 'change' estimate (along with confidence limits and an indication of significance) for each model (columns) and species (rows). These numbers are returned as a dataframe, see 'value' below. 
}
\value{
a list with three components: table - a dataframe that mimics the printed table, so in 'wide' format; estimates - a dataframe in 'long' format, one row per entry; results - a named list of the model objects, in the form "T_xxxxx_mmm", where xxxxx is the Euring code number, mmm the model specification.
}
\references{
}
\author{
Rob Robinson
}
\note{
ndigits is strict, the returned 'estimates' dataframe will be rounded to this level as well.
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{index}}
}
\examples{
#
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
