\name{extract.counts}
\alias{extract.counts}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Get species data from a CES datset
}
\description{
Extracts catch data for a particular species from a CES object, usually called through \code{extract.data}.
}
\usage{
extract.counts(data, species=0, age=0, plots=NULL)
}
\arguments{
  \item{data}{
  A dataframe containing CES data usually generated by /code{readces}
}
  \item{species}{
  Species for which to extract data. This should be a standard Euring numeric code, for example Turdus merula would be 11870 (note the last digit, to indicate race, should always be zero). If species=0, summarises catch total across all species.
}
  \item{age}{
  Euring age code 3 for juveniles and 4 for adults, any other value is an error.
}
  \item{plots}{
  An object indicating which sites were covered in each year. Normally this would be generated by \code{extract.coverage}, if it is absent the coverage data will be extracted silently.   
}
}
\details{

}
\value{
  A datatframe containing the number of birds caught at each site in each year. Usually with an identifier for the species to which the data refer.
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Rob Robinson
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{extract.data}}
}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
