\name{get.euring}
\alias{get.euring}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Find a Euring code number
}
\description{
Search for the Euring code number of a species.
}
\usage{
get.euring(string)
}
\arguments{
  \item{string}{
  A character string containing (part of) a species name, in any language.
}
}
\details{
}
\value{
  A dataframe containg the code an scientific names of species that contain the search string.
}
\references{
}
\author{
Rob Robinson
}