% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getces.R
\name{get_ces}
\alias{get_ces}
\title{Create a dataframe object from CES survey.}
\usage{
get_ces(srvy)
}
\arguments{
\item{srvy}{A CES survey code call. See \emph{Survey Code Calls} below.
\code{srvy} value must be in quotations.}
}
\description{
`get_ces()`` creates a dataframe for a requested Canadian Election Study
survey using an associated survey code to call and download
the survey dataset. Prints out the associated citation for use with
the requested dataset.
}
\details{
\subsection{Datasets}{

Datasets are loaded using either .dta or .sav file types.
To quickly convert a dataset's values to factor type use
\code{labelled::to_factor()} on the dataset.
}

\subsection{Survey Code Calls}{
\itemize{
\item ces2019_web\verb{ calls 2019 CES online survey dataset. *}ces2019_phone\verb{ calls 2019 CES phone survey dataset. *}ces2015_web\verb{ calls 2015 CES online survey dataset. *}ces2015_phone\verb{ calls 2015 CES phone survey dataset. *}ces2015_combo\verb{ calls 2015 CES combined (online/phone) dataset. *}ces2011\verb{ calls 2011 CES survey dataset. *}ces2008\verb{ calls 2008 CES survey dataset. *}ces2004\verb{ calls 2004 CES survey dataset. *}ces0411\verb{ calls 2004-2011 CES survey dataset. *}ces0406\verb{ calls 2004-2006 CES survey dataset. *}ces2000\verb{ calls 2000 CES survey dataset. *}ces1997\verb{ calls 1997 CES survey dataset. *}ces1993\verb{ calls 1993 CES survey dataset. *}ces1988` calls 1988 CES survey dataset.
\item ces1984\verb{ calls 1984 CES survey dataset. *}ces1974\verb{ calls 1974 CES survey dataset. *}ces7480\verb{ calls 1974-1980 CES survey dataset. *}ces72_jnjl` calls 1972 June-July CES survey dataset.
\item \code{ces72_sep} calls 1972 September CES survey dataset.
*\code{ces72_nov} calls 1972 November CES survey dataset.
*\code{ces1968} calls 1968 CES survey dataset.
*\code{ces1965} calls 1965 CES survey dataset.
}
}

\subsection{Incorrect/Repeated Code Calls}{

Incorrect (a non-existent survey code) or repeated (calling a code twice)
code calls will stop the function process and return associated error messages.
}

\subsection{Extra Notes}{

Due to the naming of the columns in the 1965 and 1968 datasets it is recommended
to download the associated codebook for the requested dataset.
}
}
\examples{
get_ces("ces2019_phone")

MESSAGE: TO CITE THIS SURVEY FILE: Stephenson, Laura B; Harell, Allison; Rubenson, Daniel;
Loewen, Peter John, 2020, '2019 Canadian Election Study - Phone Survey',
https://doi.org/10.7910/DVN/8RHLG1, Harvard Dataverse, V1, UNF:6:eyR28qaoYlHj9qwPWZmmVQ== [fileUNF]

get_ces("ces2005")

Error in get_ces("ces2005") : Warning: Code not in table.
}
