% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getces.R
\name{get_ces}
\alias{get_ces}
\title{Create a dataframe object from CES survey.}
\usage{
get_ces(srvy)
}
\arguments{
\item{srvy}{A CES survey code call. See \emph{Survey Code Calls} below.
\code{srvy} value must be in quotations.}
}
\description{
\code{get_ces()} creates a dataframe for a requested Canadian Election Study
survey using an associated survey code to call and download
the survey dataset. Prints out the associated citation for use with
the requested dataset.
}
\details{
\subsection{Datasets}{

Datasets are loaded using either .dta or .sav file types.
To quickly convert a dataset's values to factor type use
\code{labelled::to_factor()} on the dataset.
}

\subsection{Survey Code Calls}{
\itemize{
\item \code{ces2019_web} calls 2019 CES online survey dataset.
\item \code{ces2019_phone} calls 2019 CES phone survey dataset.
\item \code{ces2015_web} calls 2015 CES online survey dataset.
\item \code{ces2015_phone} calls 2015 CES phone survey dataset.
\item \code{ces2015_combo} calls 2015 CES combined (online/phone) dataset.
\item \code{ces2011} calls 2011 CES survey dataset.
\item \code{ces2008} calls 2008 CES survey dataset.
\item \code{ces2004} calls 2004 CES survey dataset.
\item \code{ces0411} calls 2004-2011 CES survey dataset.
\item \code{ces0406} calls 2004-2006 CES survey dataset.
\item \code{ces2000} calls 2000 CES survey dataset.
\item \code{ces1997} calls 1997 CES survey dataset.
\item \code{ces1993} calls 1993 CES survey dataset.
\item \code{ces1988} calls 1988 CES survey dataset.
\item \code{ces1984} calls 1984 CES survey dataset.
\item \code{ces1974} calls 1974 CES survey dataset.
\item \code{ces7480} calls 1974-1980 CES survey dataset.
\item \code{ces72_jnjl} calls 1972 June-July CES survey dataset.
\item \code{ces72_sep} calls 1972 September CES survey dataset.
\item \code{ces72_nov} calls 1972 November CES survey dataset.
\item \code{ces1968} calls 1968 CES survey dataset.
\item \code{ces1965} calls 1965 CES survey dataset.
}
}

\subsection{Incorrect/Repeated Code Calls}{

Incorrect (a non-existent survey code) will stop the function process and return associated error messages.
Repeated code calls will load in an unaltered version of the requested table.
}

\subsection{Extra Notes}{

Due to the naming of the columns in the 1965 and 1968 datasets it is recommended
to download the associated codebook for the requested dataset.
}
}
\examples{
get_ces("ces2019_phone")

MESSAGE: TO CITE THIS SURVEY FILE: Stephenson, Laura B; Harell, Allison; Rubenson, Daniel;
Loewen, Peter John, 2020, '2019 Canadian Election Study - Phone Survey',
https://doi.org/10.7910/DVN/8RHLG1, Harvard Dataverse, V1, UNF:6:eyR28qaoYlHj9qwPWZmmVQ== [fileUNF]

get_ces("ces2005")

Error in get_ces("ces2005") : Warning: Code not in table.
}
