% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decon.R
\name{get_decon}
\alias{get_decon}
\title{Creates a non-exhaustive dataframe of 21 variables with renamed columns.}
\usage{
get_decon()
}
\description{
get_decon() creates a non-exhaustive dataset consisting of 21 variables with renamed columns
from the demographics, ideology, and economy sections of the 2019 CES online survey.
}
\details{
NAs have not been removed, birth year (yob) has been recoded from the survey code to the actual year.
The politically left/right question variables (lr_bef and lr_aft) have also been provided in one column together under the name lr.
Variables are citizenship, yob (year of birth), gender, province_territory, education, lr, lr_bef, lr_aft,
religion, sexuality, sexuality_other (written answers), language_eng (English language speakers),
language_fr (French language speakers), language_abgl (Aboriginal language speakers), employment,
income, income cat (category), marital, econ_retro (past economic effects), econ_fed (effect of federal policies),
and econ_self (personal economic situation)
\describe{
\item{\code{citizenship}}{Canadian citizenship status}
\item{\code{yob}}{year of birth}
\item{\code{gender}}{identified gender of the respondent}
\item{\code{province_territory}}{Province or Territory of current residence}
\item{\code{education}}{highest level of education completed}
\item{\code{lr}}{united column of lr_bef and lr_aft values; whether the respondent identifies on the political spectrum}
\item{\code{lr_bef}}{where the respodent identifies on the political spectrum; asked before party identification questions}
\item{\code{lr_aft}}{where the respondent identifies on the political spectrum; asked after party identification questions}
\item{\code{religion}}{religion of respondent}
\item{\code{sexuality}}{sexual identity}
\item{\code{sexuality_other}}{sexual identity; written answers}
\item{\code{language_eng}}{language learned as child and still understand; selected response English}
\item{\code{language_fr}}{language learned as a child and still understand; selected response French}
\item{\code{language_abgl}}{langauge learned as a child and still understand; specified Aboriginal language}
\item{\code{employment}}{employment status}
\item{\code{income}}{total household income, before taxes, for the year 2018}
\item{\code{income_cat}}{selected household income category}
\item{\code{marital}}{marital status}
\item{\code{econ_retro}}{response to question, 'over the past year, has Canada's economy:'}
\item{\code{econ_fed}}{response to question, 'have the policies of the federal government made Canada's economy...'}
\item{\code{econ_self}}{response to question, have the policies of the federal government made your financial situation...'}
}
}
