\name{sitereport}
\alias{sitereport}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Site reporting summary
}
\description{
  Generate a simple summary report (as MS Word documents) of captures at a site.
}
\usage{
  sitereport(data, sitelist=NULL, year=0, template_file=NULL, outdir=getwd(), nyears=NA, verbose=TRUE)
}
\arguments{
  \item{data}{
    a dataframe generated by \code{readces}
  }
  \item{sitelist}{
    character, a vector of sitenames for which to generate the reports
  }
  \item{year}{
    specify a year to create a summary by species
  }
  \item{template_file}{
    the name of an RMarkdown template file
  }
  \item{outdir}{
    where to save the reports to
  }
  \item{nyears}{
    numeric, limit the number of years plotted to the number that the site operated, with this as a minimum. For example, if nyears=10 and the site has operated for 15 years, 15 bars (and dots) will be shown, but if the site has only operated for 5 years, then 10 bars will be plotted.
  }
  \item{verbose}{
    logical, print a message when each site has been processed?
  }
}
\details{
}
\value{
}
\note{
}
\author{
  Rob Robinson
}
\examples{
  data(ukdata)
  sitereport(ukdata, sitelist=c('656','420'))
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
